% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{canVar}
\alias{canVar}
\title{Federated ComDim}
\usage{
canVar(x_cent, value)
}
\arguments{
\item{x_cent}{centered dataset}

\item{value}{output coefficient matrix from CCA}
}
\value{
CV canonical variates
}
\description{
Federated ComDim
}
