% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{crossmatrix}
\alias{crossmatrix}
\title{Federated ComDim}
\usage{
crossmatrix(x, y = NULL)
}
\arguments{
\item{x}{first block}

\item{y}{second block}
}
\value{
crossprod between x and y and as attribute the dimension of the original dataset
}
\description{
Federated ComDim
}
