% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{hybridCrossmatrix}
\alias{hybridCrossmatrix}
\title{Federated ComDim}
\usage{
hybridCrossmatrix(x_cent, cvx)
}
\arguments{
\item{x_cent}{centered dataset}

\item{cvx}{canonical variate from x_cent dataset}
}
\value{
cv_x_cross as crossprod(cvx, x_cent)
}
\description{
Federated ComDim
}
