% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{omit}
\alias{omit}
\title{Federated ComDim}
\usage{
omit(x, nrows, Mfold)
}
\arguments{
\item{x}{dataset}

\item{Mfold}{integer stating in how many parts we want to divide the dataset}
}
\value{
omit list of rows to be omitted in k-fold cross validation
}
\description{
Federated ComDim
}
